refactor(#77/photographer): 유스케이스 위치 변경
